Cosine Law with Left arm working
