+++ body-stress-inference-python/pose.py
@@ -1,25 +1,19 @@
 import cv2
 import mediapipe as mp
-import numpy as np
 import server
 import rebaAnalysis
-import threading as th
-
+import calcUtilities as cu
 
 mp_drawing = mp.solutions.drawing_utils # this gives us all of our drawing utilities
 mp_pose = mp.solutions.pose # this is importing our pose estimation models
 cap = cv2.VideoCapture(0)
 
-def determining_joints():
-    # stage = None
-    
-    PORT  = 1755
-    # socketIsOpen = False
-    
-    # Curl reba_angle variables
-    reba_angle = 0
-    SAMPLE_SIZE = 10
+def determining_joints():    
+    PORT  = 1755    
+    reba_value = 0
+    SAMPLE_SIZE = 5
     angleArr = []
+
     with mp_pose.Pose(min_detection_confidence=0.5, min_tracking_confidence=0.5) as pose:
         while cap.isOpened():
             ret, frame = cap.read()
@@ -50,27 +44,17 @@ def determining_joints():
                 nose = [landmarks[mp_pose.PoseLandmark.NOSE.value].x,landmarks[mp_pose.PoseLandmark.NOSE.value].y]
                 right_ear = [landmarks[mp_pose.PoseLandmark.RIGHT_EAR.value].x, landmarks[mp_pose.PoseLandmark.RIGHT_EAR.value].y]
                 left_ear = [landmarks[mp_pose.PoseLandmark.LEFT_EAR.value].x, landmarks[mp_pose.PoseLandmark.LEFT_EAR.value].y]
-                
-                
-                # print("Nose - right ear: ", nose[0] - right_ear[0], 
-                #       "\nnose - left ear: ", nose[0] - left_ear[0], 
-                #       "\nright ear - left ear: ", left_ear[0] - right_ear[0],
-                #       "\n")
-
-                                
+           
                 # Calculate angle
-                # left_body_angle = calculate_angle(left_hip, left_shoulder, left_elbow)                
-                # right_body_angle = calculate_angle(right_hip, right_shoulder, right_elbow)
-                left_body_angle = calc_cosine_law(left_shoulder, left_elbow, left_hip)
-                right_body_angle = calc_cosine_law(left_hip, left_shoulder, left_elbow)
+                left_body_angle = cu.calc_cosine_law(left_shoulder, left_elbow, left_hip)
+                right_body_angle = cu.calc_cosine_law(right_shoulder, right_elbow, right_hip)
                 
                 if len(angleArr) < SAMPLE_SIZE: # take n samples and calculate average angle based off measurements
                     angleArr.append(left_body_angle)
                 else:
                     avgAngle = sum(angleArr) / len(angleArr)
                     rebaLeftArm = rebaAnalysis.CalcUpperArmPosREBA(nose[0] - left_ear[0], left_elbow[0] - left_hip[0], avgAngle) # do REBA analysis taken on angle
-                    reba_angle = rebaLeftArm # this needs to be removed
-                    print("pose angle: ", left_body_angle)
+                    reba_value = rebaLeftArm
                     angleArr = []
                     
                     body_parts = {
@@ -91,18 +75,11 @@ def determining_joints():
                 # Visualize angle
                 cv2.putText(image, str(left_body_angle), tuple(np.multiply(left_elbow, [640, 480]).astype(int)), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 2, cv2.LINE_AA)
                 cv2.putText(image, str(right_body_angle), tuple(np.multiply(right_elbow, [640, 480]).astype(int)), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 2, cv2.LINE_AA)
-                
-                # Curl reba_angle logic
-                # if angle > 160:
-                #     stage = "down"
-                # if angle < 30 and stage =='down':
-                #     stage="up"
-                #     reba_angle +=1
             except:
                 pass    
             
             # Render Reba angle
-            cv2.putText(image, 'REBA Score:' + str(reba_angle), (0,30), cv2.FONT_HERSHEY_SIMPLEX, 1, (255,255,255), 1, cv2.LINE_AA)
+            cv2.putText(image, 'REBA Score:' + str(reba_value), (0,30), cv2.FONT_HERSHEY_SIMPLEX, 1, (255,255,255), 1, cv2.LINE_AA)
                         
             # Render detections
             mp_drawing.draw_landmarks(image, results.pose_landmarks, mp_pose.POSE_CONNECTIONS,
@@ -118,30 +95,3 @@ def determining_joints():
         cv2.destroyAllWindows()
         
     return body_parts
-        
-        
-def calculate_angle(a, b, c):
-    a = np.array(a) # First
-    b = np.array(b) # Mid
-    c = np.array(c) # End
-    
-    radians = np.arctan2(c[1] - b[1], c[0] - b[0]) - np.arctan2(a[1] - b[1], a[0] - b[0])
-    angle = radians*180.0/np.pi
-    
-    if angle > 180.0:
-        angle = 360-angle
-    
-    return angle 
-
-def calc_cosine_law(a,b,c):
-    # calc_cosine_law(left_shoulder, left_elbow, left_hip)
-    a = np.array(a) # First
-    b = np.array(b) # Mid
-    c = np.array(c) # End
-    
-    abDist = np.sqrt((a[0]-b[0])**2 + (a[1]-b[1])**2)
-    acDist = np.sqrt((a[0]-c[0])**2 + (a[1]-c[1])**2)
-    bcDist = np.sqrt((b[0]-c[0])**2 + (b[1]-c[1])**2)
-    radians = np.arccos((abDist**2 + acDist**2 -bcDist**2) / (2*abDist*acDist)) 
-    angle = radians*180.0/np.pi
-    return angle
